Support plugins and plugin migrations.
To support plugins, we need to supply the path to a separate
plugins directory that can be pulled in at runtime. Plugin will
be run in order and each will supply a `process` method that will be
called. Migrations can also supply a directory of migrations. We will
merge the migrations together and run them all via utility scripts also
supplied as part of this commit.