Ensure reaped builds do not tally disk usage.
When calculating the disk usage for a project, we don't want to
factor in reaped builds. We also want to ensure that valid zeros
are returned if the project doesn't use any disk space. These
changes allow these two important cases. Tests are updated to
ensure these conditions are true.